Python_requirements.txt
tensorflow-gpu==2.9.0
pystow
unicodedata2
rdkit-pypi
jpype1
